お問い合わせフォーム完成

+++ contact-form/database/migrations/2026_05_02_141002_create_contacts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
                 ->cascadeOnDelete();
             $table->string('first_name', 255);
             $table->string('last_name', 255);
-            $table->tinyInteger('gender')
+            $table->integer('gender')
                 ->unsigned()
                 ->comment('1:男性 2:女性 3:その他');
             $table->string('email', 255);
